Update _render_text.py
It also makes sense in this case to rename the variable as it won't be the previous text's height anymore, but the current text's height. This does not change any functionality though as all text in one block should be of the same font and hence, of the same height.